runtime: storage-v2: Move `BagManager` to a separate file.
